Week 4 res2006
Hi,

Sorry this is so late.  I've been working on it a lot since last night, running bit by bit.  In the beginning I could get stuff to run up until the ones that say 'this is the title" and this is the price", but now all I can get is the map (still in black in white.)  Arggh!!!! OK, anyway, I'll keep slugging along...

Becky